+++ app/renderer/components/layout/TopBarContainer.js
@@ -1,24 +1,34 @@
 // @flow
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
+
 import TopBar from './TopBar';
 
-export default class TopBarContainer extends Component {
+const mapStateToProps = ({nodeSystem}) => ({
+  nodeSystem
+});
+
+class TopBarContainer extends Component {
 
   render() {
-    const networkName = 'Main net';
-    const isSynced = false;
+    const { nodeSystem } = this.props;
+    const { chain, name, version, isSynced } = nodeSystem;
+    const networkName = `${chain} ${version}`;
+    // const isSynced = false;
     const blockNum = 1234567890;
     const blockHeight = blockNum; // TODO
     const syncPercentage = 50;
     return (
       <TopBar
-              networkName={networkName}
-              blockNum={blockNum}
-              blockHeight={blockHeight}
-              isSynced={isSynced}
-              syncPercentage={syncPercentage}
+        networkName={networkName}
+        blockNum={blockNum}
+        blockHeight={blockHeight}
+        isSynced={isSynced}
+        syncPercentage={syncPercentage}
       />
     );
   }
 
 }
+
+export default connect(mapStateToProps)(TopBarContainer);